LSTM notebook and script
